Move ImageFileHandler to 'handlers' package
